student structure change
