+++ cgi/modules/logger.js
