+++ cgi/studentEdit.js
@@ -1,19 +1,19 @@
 const http = require('http');
 const fs = require('fs');
 const url = require('url');
-const filePathLocation = "../../data/data.json";
+const filePathLocation = "../data/data.json";
 
-http.createServer(function(req, res){
+http.createServer(function (req, res) {
     var infoFromURL = url.parse(req.url, true).query;
-    var infoNumber = infoFromURL.number;
+    var newArray = JSON.parse(infoFromURL.edit);
     try {
         if (fs.existsSync(filePathLocation)) {
             fs.readFile(filePathLocation, function(err, data) {
-                console.log("delete"); //
+                console.log("edit"); //
                 var dataArray = JSON.parse(data);
                 for (var i=0; i<dataArray.length; i++) {
-                    if (dataArray[i].uid == infoNumber) {
-                        dataArray.splice(i, 1);
+                    if (dataArray[i].uid == newArray.uid) {
+                        dataArray[i] = newArray;
                     }
                 }
                 fs.writeFileSync(filePathLocation, JSON.stringify(dataArray));
@@ -28,10 +28,7 @@ http.createServer(function(req, res){
             return res.end();
         }
     } catch (error) {
-        console.log(error);
+        console.log(error); //
     }
-}).listen(8387);
-console.log('Server started on localhost:8387; press Ctrl-C to terminate...!');
-/*
-    file:///home/givi/Desktop/DEV/ajax/image.html
-*/
+}).listen(8386);
+console.log('Server running at http://127.0.0.1:8386/');
